Added TestRun model and error messages to reports
The TestRun model stores system error messages so they can be displayed in the generated HTML report.

The nightwatch object stores all system-errs on the top level object instead of on the test case that generated the error so we show errors at the top of the page instead of on the test case that generated the error.