Added package grouping to reports.
The Default and Compact themes now show tests grouped by package.  The Cover theme still shows test results in a flat list so you can identify test failures quickly.

The styles for Default and Compact were also updated to make it easier to scan for failures.  The green text is no longer used for passing tests and instead a lighter gray is used to help them not stand out as much.

This addresses issue #6.