Added docs for --reporter option
Nightwatch has a --reporter command line option that you can use to pass the path of a reporter.  It has a slightly different interface than the normal reporter option in globals.js so users need to create another file that supports the interface to use the --reporter option.

Just added documentation for how to create the file and how to use it with the reporter option.