Fix pull request for separate filenames per suite
When I merged pull request #34 I didn't realize it made use of the old opts.reportFilename instead of the new opts.fullOutputFilename.

If the separateReportPerSuite option is passed it will add the name of the last test suite to the end of the filename.  If each of your test suites run in parallel and only one suite at a time then this would generate a unique report file per suite.